+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### next release
 - :bulb: Added list with all tracking events
 - :bulb: Added feature to track emails opening
+- :bulb: Implemented events dispatching
 
 ### v0.6.6
 - :bulb: Added feature to clean all contacts statuses